% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hechos_anio.R
\docType{data}
\name{hechos_anio}
\alias{hechos_anio}
\title{hechos_anio}
\format{
Un data frame con 150 filas y 3 variables:
\describe{
\item{anio}{Año en el que las violaciones de los derechos humanos ocurrieron.}
\item{hecho}{Tipo de violación de los derechos humanos: atentado contra el derecho a la vida, desaparición forzada, ejecución extrajudicial, privación ilegal de la libertad, tortura y violencia sexual.}
\item{violacion}{Número total de violaciones de los derechos humanos.}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(hechos_anio)
}
\description{
Datos que contienen el número total de violaciones de los  derechos humanos por año y tipo de violación en el  periodo 1984 - 2008.
}
\keyword{datasets}
