% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hechos_sexo.R
\docType{data}
\name{hechos_sexo}
\alias{hechos_sexo}
\title{hechos_sexo}
\format{
Un data frame con 12 filas y 3 variables:
\describe{
  \item{sexo}{Sexo de la víctima.}
  \item{hecho}{Tipo de violación de los derechos humanos: atentado contra el derecho a la vida, desaparición forzada, ejecución extrajudicial, privación ilegal de la libertad, tortura y violencia sexual.}
  \item{violacion}{Número total de violaciones de los derechos humanos.}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(hechos_sexo)
}
\description{
Datos que contienen el número total de violaciones de los  derechos humanos por tipo de violación y sexo de la víctima.
}
\keyword{datasets}
