% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hechos_victimas_anio.R
\docType{data}
\name{hechos_victimas_anio}
\alias{hechos_victimas_anio}
\title{hechos_victimas_anio}
\format{
Un data frame con 81 filas y 3 variables.
\describe{
\item{hecho}{Tipo de violación de los derechos humanos: atentado contra el derecho a la vida, desaparición forzada, ejecución extrajudicial, privación ilegal de la libertad, tortura y violencia sexual. }
\item{anio}{Año en el que las violaciones de los derechos humanos ocurrieron. }
\item{victima}{Número total de víctimas.}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://demografiando.pro/informe-verdad/}
}
\usage{
data(hechos_victimas_anio)
}
\description{
Datos que contienen el número total de víctimas por tipo de violación de los derechos humanos y año en Ecuador.
}
\keyword{datasets}
