% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsables.R
\docType{data}
\name{responsables}
\alias{responsables}
\title{responsables}
\format{
Un data frame con 458 filas y 2 variables.
\describe{
\item{presunto_responsable}{Nombre de los presuntos responsables de las violaciones de derechos humanos.}
\item{caso}{Casos en los que los presuntos responsables fueron señalados. }
\item{}{}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 5: Conclusiones y Recomendaciones
\url{https://www.dpe.gob.ec/}
}
\usage{
data(responsables)
}
\description{
Datos que contienen el nombre de los presuntos responsables de las violaciones de derechos humanos y el o los casos en los que fueron señalados.
}
\keyword{datasets}
