% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verdadecu-package.R
\docType{package}
\name{verdadecu-package}
\alias{verdadecu}
\alias{verdadecu-package}
\title{verdadecu: Datos de la Comision de la Verdad de Ecuador}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
\url{https://demografiando.pro/informe-verdad/}
}
\description{
El paquete \code{verdadecu} proporciona acceso a los datos recabados por la
Comisión de la Verdad de Ecuador. Permite extraer información sistematizada
para el análisis de derechos humanos en Ecuador.
}
\details{
Este paquete contiene datos sobre violaciones de derechos humanos en Ecuador
durante el período 1984-2008, incluyendo información sobre víctimas,
responsables, tipos de violaciones, ubicaciones geográficas y otros aspectos
relevantes para el análisis de derechos humanos.

Los datos provienen del Informe de la Comisión de la Verdad Ecuador 2010:
"Sin Verdad No Hay Justicia".
}
\seealso{
\url{https://github.com/Demografiando/verdadecu}
}
\author{
Adriana Robles \email{adriana@demografiando.pro}
Javier Borja \email{javier@demografiando.pro}
}
\keyword{internal}
