% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas.R
\docType{data}
\name{victimas}
\alias{victimas}
\title{victimas}
\format{
Un data frame con 456 filas y 1 variable.
\describe{
\item{victimas}{Nombre de la víctima.}
\item{}{}
\item{}{}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 5: Conclusiones y Recomendaciones
\url{https://www.dpe.gob.ec/}
}
\usage{
data(victimas)
}
\description{
Objeto que contiene el nombre de las víctimas de violaciones de derechos humanos.
}
\keyword{datasets}
