% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas_anio.R
\docType{data}
\name{victimas_anio}
\alias{victimas_anio}
\title{victimas_anio}
\format{
Un data frame con 25 filas y 2 variables.
\describe{
\item{anio}{Año en el que las violaciones de los derechos humanos ocurrieron. }
\item{victima}{Número total de víctimas.}
\item{}{}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(victimas_anio)
}
\description{
Datos que contienen el número de víctimas de violaciones de derechos humanos en Ecuador por año.
}
\keyword{datasets}
