% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas_educacion.R
\docType{data}
\name{victimas_educacion}
\alias{victimas_educacion}
\title{victimas_educacion}
\format{
Un data frame con 6 filas y 2 variables.
\describe{
\item{educacion}{Nivel educativo de la víctima.}
\item{victima}{Número total de víctimas.}
\item{}{}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(victimas_educacion)
}
\description{
Datos que contienen el número total de víctimas de violaciones de derechos humanos en Ecuador según su nivel educativo.
}
\keyword{datasets}
