% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas_nacionalidad.R
\docType{data}
\name{victimas_nacionalidad}
\alias{victimas_nacionalidad}
\title{victimas_nacionalidad}
\format{
Un data frame con 5 filas y 2 variables.
\describe{
\item{nacionalidad}{Nacionalidad de la víctima.}
\item{victima}{Número total de víctimas.}
\item{}{}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(victimas_nacionalidad)
}
\description{
Datos que contienen el número de víctimas de violaciones de derechos humanos en Ecuador según su nacionalidad.
}
\keyword{datasets}
