% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas_ocupacion.R
\docType{data}
\name{victimas_ocupacion}
\alias{victimas_ocupacion}
\title{victimas_ocupacion}
\format{
Un data frame con 13 filas y 2 variables.
\describe{
\item{ocupacion}{Ocupación de la víctima.}
\item{victima}{Número total de víctimas.}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(victimas_ocupacion)
}
\description{
Datos que contienen el número de víctimas de violaciones de derechos humanos en Ecuador según su ocupación.
}
\keyword{datasets}
