% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas_presidente.R
\docType{data}
\name{victimas_presidente}
\alias{victimas_presidente}
\title{victimas_presidente}
\format{
Un data frame con 11 filas y 2 variables.
\describe{
\item{presidente}{Presidente de la República del periodo correspondiente en el cual ocurrieron las violaciones de derechos humanos. }
\item{victima}{Número total de víctimas.}
\item{}{}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(victimas_presidente)
}
\description{
Datos que contienen el número de víctimas según el periodo presidencial en el cual ocurrieron las violaciones de derechos humanos.
}
\keyword{datasets}
