% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victimas_provincia.R
\docType{data}
\name{victimas_provincia}
\alias{victimas_provincia}
\title{victimas_provincia}
\format{
Un data frame con 29 filas y 3 variables.
\describe{
\item{provincia}{Provincia en el que se produjeron las violaciones de derechos humanos.}
\item{ciudad}{Ciudad en el que se produjeron las violaciones de derechos humanos.}
\item{victima}{Número total de víctimas.}
}
}
\source{
Informe de la Comisión de la Verdad Ecuador 2010. Sin Verdad No Hay Justicia.
Tomo 1: Violaciones de los Derechos Humanos.
\url{https://www.dpe.gob.ec/}
}
\usage{
data(victimas_provincia)
}
\description{
Datos que contienen el número de víctimas según el lugar en el que se produjeron las violaciones de derechos humanos.
}
\keyword{datasets}
